+++ src/3_game_loop/a_game_logic/entities/orb_entity.c
@@ -15,28 +15,30 @@
 void	tick_orb(t_entity *self)
 {
 	t_orb	*orb;
+	int		ammo_amount;
 
 	if (pos_distance(player()->transform.pos, self->transform.pos) > GOAL_COLLISION_DISTANCE)
 		return ;
+	ammo_amount = random_int(MIN_AMMO_AMOUNT_PER_ORB, MAX_AMMO_AMOUNT_PER_ORB);
 	orb = (t_orb *)self->data;
 	if (orb->type == TYPE_EARTH)
 	{
-		player()->inv.earth_orbs += AMMO_AMOUNT_PER_ORB;
+		player()->inv.earth_orbs += ammo_amount;
 		logger(LOGGER_ACTION, "Picked up some earth orbs!");
 	}
 	else if (orb->type == TYPE_FIRE)
 	{
-		player()->inv.fire_orbs += AMMO_AMOUNT_PER_ORB;
+		player()->inv.fire_orbs += ammo_amount;
 		logger(LOGGER_ACTION, "Picked up some fire orbs!");
 	}
 	else if (orb->type == TYPE_WATER)
 	{
-		player()->inv.water_orbs += AMMO_AMOUNT_PER_ORB;
+		player()->inv.water_orbs += ammo_amount;
 		logger(LOGGER_ACTION, "Picked up some water orbs!");
 	}
 	else if (orb->type == TYPE_AIR)
 	{
-		player()->inv.air_orbs += AMMO_AMOUNT_PER_ORB;
+		player()->inv.air_orbs += ammo_amount;
 		logger(LOGGER_ACTION, "Picked up some air orbs!");
 	}
 	self->to_be_deleted = true;